requirements.txt
nose
djangorestframework
markdown
django-filter